+++ cmd/tempo-cli/cmd-list-index.go
@@ -33,6 +33,10 @@ func (cmd *listIndexCmd) Run(ctx *globalOptions) error {
 		return err
 	}
 
+	if meta.Version != v2.VersionString {
+		return fmt.Errorf("unsupported block version: %s", meta.Version)
+	}
+
 	b, err := v2.NewBackendBlock(meta, r)
 	if err != nil {
 		return err