+++ tempodb/encoding/v2/iterator_record.go
@@ -15,14 +15,14 @@ type recordIterator struct {
 	objectRW common.ObjectReaderWriter
 	dataR    common.DataReader
 
-	currentIterator Iterator
+	currentIterator common.Iterator
 
 	buffer []byte
 }
 
 // NewRecordIterator returns a recordIterator.  This iterator is used for iterating through
 //  a series of objects by reading them one at a time from Records.
-func NewRecordIterator(r []common.Record, dataR common.DataReader, objectRW common.ObjectReaderWriter) Iterator {
+func NewRecordIterator(r []common.Record, dataR common.DataReader, objectRW common.ObjectReaderWriter) common.Iterator {
 	return &recordIterator{
 		records:  r,
 		objectRW: objectRW,